Add deployment script.
The deployment script is written to clone and build each version of
Katello automatically based on the branch names for releases
(e.g. KATELLO-2.0). The result of the script is the fully built (html)
version of master/nightly combined with the docs for each version built
with the current nightly layouts and assets and placed under docs/<version>